% Part of the oro.dicom package for R
% Distributed under the BSD license: see oro.dicom/COPYING
%
% $Id: $

\name{DICOM Dictionary}
\alias{dicom.dic.thompson}
\alias{dicom.dic}
\alias{dicom.VR}
\title{Lookup Tables for DICOM Header Information}
\description{
  Lookup tables for DICOM header information.
}
\usage{
data(dicom.dic.thompson)
data(dicom.dic)
data(dicom.VR)
}
% \format{}
\source{
  See references.
}
\references{
  Digital Imaging and Communications in Medicine (DICOM)\cr
  \url{http://medical.nema.org}\cr
  \url{http://en.wikipedia.org/wiki/Digital_Imaging_and_Communications_in_Medicine}
}
\seealso{
  \code{\link{readDICOM}}, \code{\link{readDICOMFile}}.
}
\keyword{datasets}
