% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oro.dicom-package.R
\docType{data}
\name{dicom.dic}
\alias{dicom.VR}
\alias{dicom.dic}
\alias{dicom.dic.thompson}
\title{Lookup Tables for DICOM Header Information}
\format{\preformatted{'data.frame':	4188 obs. of  5 variables:
 $ group  : chr  "0000" "0000" "0000" "0000" ...
 $ element: chr  "0000" "0001" "0002" "0003" ...
 $ code   : chr  "UL" "UL" "UI" "UI" ...
 $ offset : chr  "1" "1" "1" "1" ...
 $ name   : chr  "GroupLength" "CommandLengthToEnd" "AffectedSOPClassUID" "RequestedSOPClassUID" ...
}}
\source{
See references.
}
\usage{
data(dicom.dic)
}
\description{
Lookup Tables for DICOM Header Information
}
\references{
Digital Imaging and Communications in Medicine (DICOM)\cr
\url{http://medical.nema.org}\cr
\url{http://en.wikipedia.org/wiki/Digital_Imaging_and_Communications_in_Medicine}
}
\seealso{
\code{\link{readDICOM}}, \code{\link{readDICOMFile}}.
}
\keyword{datasets}

