% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oro.dicom-package.R
\docType{data}
\name{dicom.dic}
\alias{dicom.dic}
\alias{dicom.dic.thompson}
\alias{dicom.VR}
\title{Lookup Tables for DICOM Header Information}
\format{An object of class \code{data.frame} with 4188 rows and 5 columns.}
\source{
See references.
}
\usage{
data(dicom.dic)
}
\description{
Lookup Tables for DICOM Header Information
}
\references{
Digital Imaging and Communications in Medicine (DICOM)\cr
\url{http://medical.nema.org}\cr
\url{http://en.wikipedia.org/wiki/Digital_Imaging_and_Communications_in_Medicine}
}
\seealso{
\code{\link{readDICOM}}, \code{\link{readDICOMFile}}.
}
\keyword{datasets}
