\name{DICOM Lookup Tables}
\alias{dicom.fields.data}
\title{Lookup Tables for DICOM Header Information}
\description{
  Lookup tables for DICOM header information.
}
\usage{
data(dicom.fields.data)
}
% \format{}
\source{
  See references.
}
\references{
  Digital Imaging and Communications in Medicine (DICOM)
  \url{http://medical.nema.org}
  \url{http://en.wikipedia.org/wiki/Digital_Imaging_and_Communications_in_Medicine}
}
\seealso{
  \code{\link{dicomInfo}}, \code{\link{dicomSeparate}}.
}
\keyword{datasets}
