\name{dicomSeparate}
\alias{dicomSeparate}
\title{Read All DICOM Files in a Directory}
\description{
  All DICOM files are imported and a text file summarizing their content
  recorded.
}
\usage{
dicomSeparate(path, debug=TRUE)
}
\arguments{
  \item{path}{Path name to the DICOM directory.}
  \item{debug}{Flag to provide limited progress information.}
}
\details{
}
\value{
}
\references{
  Digital Imaging and Communications in Medicine (DICOM)\cr
  \url{http://medical.nema.org}
}
\seealso{
  \code{\link{dicomInfo}}
}
%\examples{}
\author{Brandon Whitcher \email{bjw34032@users.sourceforge.net}}
\keyword{file}
