% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header2matrix}
\alias{header2matrix}
\title{Converts DICOM Header Field to a Matrix}
\usage{
header2matrix(hdr, ncol, sep = " ", byrow = TRUE)
}
\arguments{
\item{hdr}{is the result from extracting information from a DICOM header
field; e.g., using \code{extractHeader}.}

\item{ncol}{is the number of columns.}

\item{sep}{is the character string required to split entries in the header
field.}

\item{byrow}{is a logical variable (default = \code{TRUE}) telling the
routine to populate the matrix by rows then columns.}
}
\value{
Matrix with \code{length(hdr)} rows and \code{ncol} columns.
}
\description{
Converts a vector of DICOM header information, assuming there are multiple
entries per element of the vector, into a matrix.
}
\examples{

x <- readDICOMFile(system.file("dcm/Abdo.dcm", package="oro.dicom"))
pixelSpacing <- extractHeader(x$hdr, "PixelSpacing", numeric=FALSE)
pSmat <- header2matrix(pixelSpacing, ncol=2)
IOP <- extractHeader(x$hdr, "ImageOrientationPatient", numeric=FALSE)
IOPmat <- header2matrix(IOP, ncol=6)

}
\references{
Digital Imaging and Communications in Medicine (DICOM)\cr
\url{http://medical.nema.org}
}
\seealso{
\code{\link{extractHeader}}, \code{\link{matrix}}
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\keyword{misc}
