% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/header.R
\name{writeHeader}
\alias{writeHeader}
\title{Write DICOM Table to ASCII File}
\usage{
writeHeader(dtable, filename, ...)
}
\arguments{
\item{dtable}{The DICOM table.}

\item{filename}{Name of the file to be created.}

\item{...}{Additional parameters to be passed to \code{write.table}.}
}
\value{
None.
}
\description{
A wrapper to \code{write.table} specifically for DICOM tables.
}
\details{
This function is a straightforward wrapper to \code{write.table}.
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
Digital Imaging and Communications in Medicine (DICOM)\cr
\url{http://medical.nema.org}
}
\seealso{
\code{\link{write.table}}
}
\keyword{file}

